Changed how query and menu cancel and finish routines destroy the region.
darcs-hash:20050305124204-4eba8-fe8b932a943bf6acf5135780f85a0548f06c228f.gz

+++ mod_query/wedln.c
@@ -630,7 +630,8 @@ static void wedln_do_finish(WEdln *wedln)
     wedln->handler=extl_fn_none();
     p=edln_finish(&(wedln->edln));
     
-    destroy_obj((Obj*)wedln);
+    if(region_manager_allows_destroying((WRegion*)wedln))
+       destroy_obj((Obj*)wedln);
     
     if(p!=NULL)
         extl_call(handler, "s", NULL, p);
